+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Notable changes of tdmclient-ty. Release versions refer to [https://pypi.org/pro
 - Shortcut Control-Shift-T for menu command Tools>Transpile Program
 - Menu commands moved to a separate "Thymio" menu
 - Menu entry "Disconnect" disabled when the robot is not connected
+- In error messages, hyperlink to select line
 
 ### Fixed